GafferSceneUI : fix utf-8 errors on Windows
There seems to be an encoding mismatch when importing Python files with certain utf-8 characters where the files are assumed to have cp1252 encoding which can't convert some characters to utf-8. Even forcing the file to be saved as utf-8 does not fix the problem. Replacing "real" utf-8 characters (wysiwyg in a text editor) with the equivalent utf-8 code solves this.