FileSystemPathTest : Improve file ownership query
The previous scheme of using the `USERNAME` environment variable does not work in at least one specific case, which is used by the Github Workflows systems. When creating a file as a user that is part of the `Administrators` group, Windows will actually set the _owner_ of that file to `Administrators` (which is actually a group). This fails the test that assumes it would be `USERNAME`.