WidthHeightHandle : Add `HandleProperties` struct
The constructor argument list was starting to get pretty long, and there
are more parameters coming to support other light types. This makes
it a bit more readable and compact.

I also moved to using a full `V3f` axis for the width and height
axes instead of a single float. This will be needed by other light
types.