CI : Remove quotes around `test` arguments
This is needed to run the Arnold tests on Windows. Because we are using
`subprocess` to run those tests, we're running via `cmd.exe` instead of
Powershell (as with the rest of the tests). `cmd.exe` doesn't have
`Invoke-Expression`, so we remove that from the Windows test runner
and just let everything run without the quoted command.